Refactoring: moving multiple is_frozen definitions to tribler_common
