+++ src/tribler-common/tribler_common/utilities.py
@@ -1,0 +1,13 @@
+import sys
+
+
+def is_frozen():
+    """
+    Return whether we are running in a frozen environment
+    """
+    try:
+        # PyInstaller creates a temp folder and stores path in _MEIPASS
+        sys._MEIPASS
+    except Exception:
+        return False
+    return True
